requirements.txt
torch==1.13.1
torchvision==0.14.1
matplotlib==3.7.0
Pillow==9.4.0
opencv-python==4.5.5.64
numpy==1.21.6